backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
python-multipart==0.0.6
pandas==2.2.0
openpyxl==3.1.2
numpy==1.26.3
pydantic==2.5.3
python-dotenv==1.0.0
cryptography==42.0.0
